+++ src/main/java/domain/card/Cards.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.card;
+
+import domain.score.Score;
 
 import java.util.ArrayList;
 import java.util.List;
