+++ src/main/java/domain/card/Denomination.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public enum Denomination {
