+++ src/main/java/domain/player/Dealer.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 import java.util.List;
 import java.util.stream.Collectors;
