+++ src/main/java/domain/player/DealerStatus.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 public enum DealerStatus {
     WIN,
