+++ src/main/java/domain/player/PlayerName.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 public class PlayerName {
