+++ src/main/java/view/InputView.java
@@ -1,6 +1,6 @@
 package view;
 
-import domain.Player;
+import domain.player.Player;
 
 import java.util.Arrays;
 import java.util.List;
