+++ src/test/java/domain/PlayerNameTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.player.PlayerName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.params.ParameterizedTest;
